refactors Training component
