+++ src/components/Pokemon/Breeding/index.tsx
@@ -3,9 +3,7 @@ import type { PokemonSpecies, EvolutionChain } from 'pokenode-ts';
 // components
 import Box, { BoxProps } from '@/components/Box';
 // helpers
-import { AnimatePresence } from 'framer-motion';
 import { removeDash } from '@/helpers/typography';
-import { fadeInUpVariant } from '@/helpers/animations';
 // styles
 import { SectionTitle, Table, Numbered } from '@/components/BaseStyles';
 
@@ -38,37 +36,30 @@ const Breeding = ({ species, evolutionChain, ...rest }: BreedingProps): JSX.Elem
   return (
     <Box align={{ xxs: 'center', lg: 'flex-start' }} {...rest}>
       <SectionTitle>Breeding</SectionTitle>
-      <AnimatePresence>
-        <Table
-          initial="hidden"
-          animate="show"
-          variants={fadeInUpVariant}
-          key="pokemon-breeding-table"
-        >
-          <tbody>
-            <tr>
-              <th>Gender Distribution</th>
-              <td>{gender_rate === -1 ? 'Genderless' : genderRatio(gender_rate)}</td>
-            </tr>
-            <tr>
-              <th>Egg Groups</th>
-              <td>{egg_groups.length ? eggGroups(egg_groups) : 'No Egg Groups'}</td>
-            </tr>
-            <tr>
-              <th>Egg Cycles</th>
-              <td>{hatch_counter ? eggCycle(hatch_counter) : 'No Egg Cycles'}</td>
-            </tr>
-            <tr>
-              <th>Baby Trigger Item</th>
-              <td>{baby_trigger_item ? removeDash(baby_trigger_item.name) : 'None'}</td>
-            </tr>
-            <tr>
-              <th>Habitat</th>
-              <td>{habitat ? removeDash(habitat.name) : 'None'}</td>
-            </tr>
-          </tbody>
-        </Table>
-      </AnimatePresence>
+      <Table>
+        <tbody>
+          <tr>
+            <th>Gender Distribution</th>
+            <td>{gender_rate === -1 ? 'Genderless' : genderRatio(gender_rate)}</td>
+          </tr>
+          <tr>
+            <th>Egg Groups</th>
+            <td>{egg_groups.length ? eggGroups(egg_groups) : 'No Egg Groups'}</td>
+          </tr>
+          <tr>
+            <th>Egg Cycles</th>
+            <td>{hatch_counter ? eggCycle(hatch_counter) : 'No Egg Cycles'}</td>
+          </tr>
+          <tr>
+            <th>Baby Trigger Item</th>
+            <td>{baby_trigger_item ? removeDash(baby_trigger_item.name) : 'None'}</td>
+          </tr>
+          <tr>
+            <th>Habitat</th>
+            <td>{habitat ? removeDash(habitat.name) : 'None'}</td>
+          </tr>
+        </tbody>
+      </Table>
     </Box>
   );
 };
