+++ src/components/Pokemon/index.tsx
@@ -155,10 +155,12 @@ const PokemonPage = ({
             margin={{ xxs: '0 0 2rem', lg: '0' }}
             padding={{ xxs: '0', lg: '0 2rem 0 0' }}
           />
-          {/* <Training
-                margin={{ xxs: '0 0 2rem', lg: '0' }}
-                padding={{ xxs: '0', lg: '0 1rem' }}
-              /> */}
+          <Training
+            pokemon={pokemon}
+            species={species}
+            margin={{ xxs: '0 0 2rem', lg: '0' }}
+            padding={{ xxs: '0', lg: '0 1rem' }}
+          />
           {/* <Multipliers
                 margin={{ xxs: '0 0 2rem', lg: '0' }}
                 padding={{ xxs: '0', lg: '0 0 0 2rem' }}
